+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features (ENH)
 - New `extract-annotated-pages` to filter out only the user annotated pages ([PR #98](https://github.com/py-pdf/pdfly/pull/98))
 - Added optional `--password` argument to `cat` to perform decryption ([PR #61](https://github.com/py-pdf/pdfly/pull/61))
+- `pagemeta` now display known page formats when it can detect it: A3, A4, A5, Letter, Legal
 
 ### Bug Fixes (BUG)
 - `pypdf[full]` is now a dependency, instead of just `pypdf`, to avoid some cases of `DependencyError`